+++ src/pipeline/Middleware.js
@@ -1,3 +1,5 @@
+import Pipeline from './Pipeline'
+
 export default class Middleware {
   constructor (next) {
     this.next = next
@@ -5,5 +7,10 @@ export default class Middleware {
 
   pipe (request) {
     return this.next(request)
+  }
+
+  embed (...middleware) {
+    return Pipeline.make(middleware, this.next.__decorators || [])
+      .then(this.next)
   }
 }