+++ src/server/middleware/CompressionMiddleware.js
@@ -1,18 +1,24 @@
 import GzipMiddleware from './GzipMiddleware'
 import DeflateMiddleware from './DeflateMiddleware'
+import Middleware from '../../pipeline/Middleware'
 
-export default function CompressionMiddleware (next) {
-  const gzip = new GzipMiddleware(next)
-  const deflate = new DeflateMiddleware(next)
-  return (request) => {
-    if (gzip.canEncode(request)) {
-      return gzip.pipe(request)
+export default class CompressionMiddleware extends Middleware {
+  constructor (next) {
+    super(next)
+
+    this.gzip = this.embed(GzipMiddleware)
+    this.deflate = this.embed(DeflateMiddleware)
+  }
+
+  pipe (request) {
+    if (GzipMiddleware.canEncode(request)) {
+      return this.gzip.pipe(request)
     }
 
-    if (deflate.canEncode(request)) {
-      return deflate.pipe(request)
+    if (DeflateMiddleware.canEncode(request)) {
+      return this.deflate.pipe(request)
     }
 
-    return next(request)
+    return this.next(request)
   }
 }